requirements.txt
# ============================
# Framework principal (API)
# ============================
Flask==3.0.3
Flask-Cors==4.0.1

# ============================
# Banco de Dados
# ============================
pg8000==1.31.2

supabase==1.0.3


python-dotenv==1.0.1

# ============================
# IA • Machine Learning
# ============================
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
scipy==1.12.0

# ============================
# Utilidades gerais
# ============================
requests==2.31.0

# ============================
# Segurança (hash de senhas)
# ============================
bcrypt==4.1.2

# ============================
# Visualização (se quiser gráficos no backend)
# ============================
matplotlib==3.8.2

# ============================
# Logs e ferramentas internas
# ============================
coloredlogs==15.0.1

# ============================
# Organização de datas (IA/Planner)
# ============================
python-dateutil==2.8.2

# ============================
# Para evitar warnings 
# ============================
joblib==1.3.2
